+++ src/ios/SimpleCameraPreview.h
@@ -10,7 +10,6 @@
     CLLocation* currentLocation;
 }
 
-- (void) setOptions:(CDVInvokedUrlCommand*)command;
 - (void) enable:(CDVInvokedUrlCommand*)command;
 - (void) disable:(CDVInvokedUrlCommand*)command;
 - (void) capture:(CDVInvokedUrlCommand*)command;