+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [2.0.40](https://github.com/spoonconsulting/cordova-plugin-simple-camera-preview/compare/2.0.39...2.0.40) (2025-05-23)
+* **Android:** Add Aspect Ratio option
+* **iOS:** Add Aspect Ratio option
+
 ## [2.0.39](https://github.com/spoonconsulting/cordova-plugin-simple-camera-preview/compare/2.0.38...2.0.39) (2025-05-06)
 * **Android:** Enhance torch error reporting